+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/selection/OneItemDetail.kt
@@ -14,8 +14,9 @@ internal class OneItemDetail(
     override fun getPosition(): Int = adapterPosition
     @Nullable override fun getSelectionKey(): Long = selectionKey
 
+    // or: should any click trigger selection?
     override fun inSelectionHotspot(e: MotionEvent): Boolean = when (selectionTrigger) {
         SelectionStateConfig.SelectionTrigger.Click -> true
-        SelectionStateConfig.SelectionTrigger.LongClick -> super.inSelectionHotspot(e)
+        SelectionStateConfig.SelectionTrigger.LongClick, SelectionStateConfig.SelectionTrigger.Manual -> false
     }
 }